requirements.txt
langchain==0.0.174
python-dotenv==1.0.0
